% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keyword.R
\name{kw_scatter}
\alias{kw_scatter}
\title{Plot a keyword-filtered scatter graph}
\usage{
kw_scatter(
  data,
  keyword,
  keyword_column,
  x,
  y,
  xlogscale = FALSE,
  ylogscale = FALSE,
  xtrim = 0,
  ytrim = 0,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{data}{The tibble on which the function is being performed.}

\item{keyword}{The string that the keyword_column will be filtered for.}

\item{keyword_column}{The column that will be filtered to only include entries that contain the keyword. Input should be a character / factor column.}

\item{x}{The column to use for the x-axis of the plot. Input should be a column that can be used for a scatter graph (eg numeric, datetime).}

\item{y}{The column to use for the y-axis of the plot. Input should be a column that can be used for a scatter graph (eg numeric, datetime).}

\item{xlogscale}{Option to use a log-scale on the x-axis.}

\item{ylogscale}{Option to use a log-scale on the y-axis.}

\item{xtrim}{Option to trim a certain proportion of the x-axis data (split across either end of the value range). Value should be between 0 and 1.}

\item{ytrim}{Option to trim a certain proportion of the y-axis data (split across either end of the value range). Value should be between 0 and 1}

\item{na.rm}{Choice of keeping or removing NA values.}

\item{...}{Further arguments to be used inside the geom_point() function of the plot (eg colour = "Red")}
}
\value{
A ggplot geom_point() graph
}
\description{
Plot a keyword-filtered scatter graph
}
\examples{

kw_scatter(data = gapminder::gapminder, keyword = "United", keyword_column = country, x = lifeExp, y = gdpPercap)

}
